SdioSdCard: Add a 1ms delay between turning on the SD clock and sending CMD0.
This fixes initialization problems with certain cards.
The SD spec requires providing at least 74 clocks to the card before sending CMD0. On the STM32, this amounts to 74/(400KHz)=0.185ms.
Make that a conservative 1ms so hopefully we don't ever have to ask the SDIO driver about the clock frequency used for initialization.