AxisDriver: Implement an command pre-loading.
This changes the behavior such that a new command is loaded when the last step for the previous command is issued, not when the first step for the new command should be issued. This is useful when the stepper driver requires the DIR signal to precede the the STEP signal a larger amount of time. Without this (assuming step delays are enabled) we would see a lot of wasteful waiting in the interrupts at the start of stepper commands.