Implement a T parameter for setting the max speed as nominal move time.
This allows specifying the maximum speed independent of a particular definition of distance.
The maximum speed will be the speed at which the move would be performed if took exactly T seconds at constant speed. In reality it will take at least T time but may take longer due to accelerations and axis speed limits.
In other words, the segment-relative speed will be limited to 1/T [segments per second].