AxisDriver: Improve memory usage on ARM.
Instead of having the pre-left-shifted acceleration value in the Command structures, have the original value there, and do the necessary shifting when the command is loaded.
This way the acceleration field in Command is 16 bits wide, not 32.
We now support both ways, because the old way is still optimal on AVR.