Restucturing the configuration with respect to steppers associated with an axis.
Now for each "axis" we have a set of "steppers". There is no primary stepper any more, all are equivalent and configured as a list.
The only exception is that the timer unit for the AxisDriver is taken from the stepper port for the first stepper of the axis.

We also now support motor current and microstep configuration for each stepper of an axis.

The GUI will automatically port old configurations.