FatFs, BlockCache: Respect eviction priorities completely.
When allocating a block, previously if there was an immediately reassignable block available, we took it. Now we may delay the block allocation and push for release if we have unreferenced dirty blocks with a lesser eviction priority number.
This should reliably prevent data blocks from causing repeated eviction of metadata blocks that could otherwise have been reused.