HttpServer, WebInterfaceModule: Simplify and improve timeout logic.
Get rid of the magic combined-timeout which didn't really work properly with both request and response were being processed (it never timed out).
Instead the user not tells HttpServer when they are (not) waiting for RX or TX. They can also immediately cause a disconnect; this is used in the gcode send-buf timeout.