lwip: Remove the memory alignment hacking in favor of structs and unions.
Note, we no longer do magic alignment of pbuf ->payload in pbuf_alloc. The alignment of payloads in pbufs with data in pools can be controlled using PBUF_PAYLOAD_ALIGN_TYPE though but this affects the raw start only, before any headers. This should be fine because packed structs are used for protocol headers anyway.