lwip: Refactoring of implementation and usage of pbuf_header.
Now we have:
- pbuf_header: Can only reveal headers, and checks for available space always.
- pbuf_header_force: Can only reveal headers, does not check for space (actually, if it can be checked, asserts that it's available). Changed to return void.
- pbuf_unheader: New function for hiding headers. Returns void since callers may only call if the adjustment is possible.

Return values are changed to follow the principle that functions only return errors for conditions which are permitted according to the API. Violations of the API are checked with assertions only and do not have associated error codes. If there is no error a function can encounter beyond incorrect API usage, it shall return void.

Also some refactoring and as small fix is done in tcp_in.c related to skipping over the TCP options. See: https://savannah.nongnu.org/bugs/?48402