TCP: Use a union for TcpConnection and TcpListener pointers.
Currently a flag LIS_LINK is used to distinguish the case where a connection has just entered ESTABLISHED but not yet been accepted by the application. This means lots of care and assertions are needed to ensure access to ->con only when LIS_LINK is not set. Possibly this can be improved by delaying transition to ESTABLISHED until the user accepts the connection.