Get rid of the LIS_LINK flag.
Also protect a PCB which is just being accepted from being aborted by allocate_pcb(). That would be bad because if the application tries to accept after that, a crash would result. This protection is done by removing the PCB from the list of unreferenced PCBs just before the connectionAccepted callback.