TCP: Improve handling of send errors.
If sending a segment fails in pcb_output_queued, stop the loop and effectively don't assume that the segment was sent. Also use the OutputTimer to retry sending after some time, and if the error was the send buffer being full, retry much quicker than otherwise.

This seems to work better because the unsent segment doesn't need to be retransmitted only after we figure out it was lost. Actually it is very likely it will be transmitted much sooner than the timeout, upon the next ACK or when more data is queued.