TCP: Using just one real timer for each PCB.
This reduces size of TCP PCBs significantly. The new MultiTimer implements multiplexing multiple timers onto one real timer.

However, it does seem to reduce performance for about 1%. This is not a big
concern currently, the memory usage reduction is much greater in comparison.
But a possible future direction to look into is a timers heap for all TCP PCBs.