Major build system refactoring.
The generate.py script no longer outputs a Nix expression but pure data which is an input to a Nix expression.
The build is done by building the aprinterBuild package (implemented in aprinter.nix) with a specific config
file, and that will internally run generate.py and use its output to do the right things. generate.py is changed
to have more responsibility for providing low-level build instructions like compile flags. The board definitions
(boards.nix) are gone.

For convenience, tools/build.py is created to simplify building. It is a simple Python script that exec's
nix-build with the right options.